Images are now stock in FileSystem

+++ Helloworld/CoreDataGif.swift
@@ -12,11 +12,11 @@ import CoreData
 
 class CoreDataGif {
    
-    func add(gifGiphy: GifGiphy) -> Gif? {
+    func add(gifGiphy: GifGiphy) -> Bool {
         
         guard let appDelegate =
             UIApplication.shared.delegate as? AppDelegate else {
-                return nil
+                return false
         }
         
         // 1
@@ -38,17 +38,32 @@ class CoreDataGif {
         
         gif.id = gifGiphy.id
         gif.minUrl = gifGiphy.minUrl
-        gif.minImage = gifGiphy.minNSDataImage
+        //gif.minImage = gifGiphy.minNSDataImage
         gif.originUrl = gifGiphy.originUrl
         
+        if let documentDirectoryURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first {
+            let imagePath = documentDirectoryURL.appendingPathComponent("\(gif.id!)_min")
+            do {
+                guard let minImage = gifGiphy.minNSDataImage else {
+                    print("gif.minImage is nil")
+                    return false
+                }
+                try minImage.write(to: imagePath)
+                print("Saving: \(imagePath)")
+            } catch {
+                fatalError("Can't write image \(error)")
+            }
+        }
         
         // 4
         do {
             try managedContext.save()
-            return gif
+            gifGiphy.gif = gif
+            gifGiphy.loadGIF()
+            return true
         } catch let error as NSError {
             print("Could not save. \(error), \(error.userInfo)")
-            return nil
+            return false
         }
     }
     
@@ -56,16 +71,43 @@ class CoreDataGif {
         guard let context = gif.managedObjectContext else {
             return false
         }
-        context.delete(gif)
         do {
-            try context.save()
+            if let documentDirectoryURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first {
+                var imagePath = documentDirectoryURL.appendingPathComponent("\(gif.id!)_min")
+                
+                if FileManager.default.fileExists(atPath: imagePath.path) {
+                    do {
+                        try FileManager.default.removeItem(at: imagePath)
+                        print("Delete: \(imagePath)")
+                    } catch {
+                        print("Can't delete image \(error)")
+                    }
+                }
+                else {
+                    print("Can't delete minImage: \(imagePath)")
+                }
+                imagePath = documentDirectoryURL.appendingPathComponent("\(gif.id!)_max")
+                if FileManager.default.fileExists(atPath: imagePath.path) {
+                    do {
+                        try FileManager.default.removeItem(at: imagePath)
+                        print("Delete: \(imagePath)")
+                    } catch {
+                        print("Can't delete image \(error)")
+                    }
+                }
+                else {
+                    print("Can't delete originImage: \(imagePath)")
+                }
+                context.delete(gif)
+                try context.save()
+            }
             return true
         } catch let error as NSError {
             print("deleteGif Error: \(error)")
             return false
         }
     }
-    
+
     func fetchGif() -> [Gif]?{
         
         //1
